Add retry to testing harness (#33)
Some commands in the testing harness fail sometimes e.g. pulling LXD
images because of network problems.
This commit adds a `run_with_retry` method to avoid this kind of
flakyness in the CI and applies it to the LXD launch command (where we
have seen these network problem already).