Add snap module
The snap module provides helpers for common interactions with the snap,
e.g. to start a service or create a snap path.

Existing use of snap environment variables was updated.

Add k8s-dqlite clustering support

This commit integrates the clustering of k8s-dqlite into the REST-API.
For that, the k8s-dqlite certificate will be written into k8sd when
initializing k8s. When a new node joins the cluster it will first join
k8sd and then use the certificate and key from the database to join the
k8s-dqlite cluster.