Move k8s init behind REST API
To bootstrap and configure the k8s cluster without the CLI (e.g. from
the Charm) we need to be able to initialize the cluster via REST API.

This commit moves the init steps for bootstrapping the cluster to the
`POST /cluster` endpoint and adds setup steps for joining nodes to the
`POST /cluster/{node}` endpoint.