Explicitly set timeout in each CLI command
Right now, the context deadline is initially set at root level when the command is created.
This causes problems for interactive commands as the deadline may be exceeded while the user inputs the data.
To prevent this, we remove the global timeout and specifically set it in
each command right before the RPC calls.