Forces volume umount when snap is removed
Some Kubernetes volumes may not be easily removed (e.g.: NFS volumes
from a local NFS provider), leading to a significantly increased snap
removal time and leaked volume mounts.

Due to the volume leak, the /var/lib/kubelet folder is never cleaned up
properly either, which means that on reinstalling the k8s snap
afterwards and bootstrapping the cluster, the current node will not be
registered to the cluster.

This adds the -f option to umount, forcing the unmount.