Removes lingering kubelet plugins on snap removal
At the moment, the kubelet plugins in /var/lib/kubelet/plugins/ are not
removed, even though the Pods associated with those plugins have been
removed.

This means that on reinstallation, kubelet will try to access those
inexistent plugins through the *.sock files, including for plugins we
don't want anymore, and delaying registration for the ones we do want.