Add cert refresh e2e test
We're adding an integration test that refreshes the certificates
for control plane nodes as well as worker nodes. It then verifies
that the new certificates have the requested expiry date and
additional Subject Alternative Names.

For simplicity, we'll configure the cluster to automatically
approve CSR requests.

While at it, we're adding E231 and E226 to the flake8 ignore list
since the linter cannot properly handle format strings.