Fix flaky node remove tests
k8sd sometimes fails when requested to remove nodes immediately
after bootstrapping the cluster. It seems that it takes a little
longer for trust store changes to be propagated to all nodes, which
should probably be fixed on the microcluster side.

For now, we'll perform some retries.

    failed to POST /k8sd/cluster/remove: failed to delete cluster member
    k8s-integration-c1aee0-2: No truststore entry found for node with name
    "k8s-integration-c1aee0-2"

While at it, we're adding some flake8 rules to the ignore list
since the linter cannot properly handle formatted strings.