Reduce the scope of "session_instance"
"session_instance" is a fixture that allows sharing the same k8s
environment for multiple tests, aiming to reduce test duration.

The problem is that some tests interfere with each other and fail.

We could try to improve the test cleanup, ensuring that the shared
environment is brought back to a "clean" state when the test
finishes, however we'd most probably continue to have flaky tests.

To address this, we're reducing the scope of "session_instance" to
function scope. Since it provides a single node cluster with all
features enabled, we'll rename it to "aio_instance".

Isolating the tests may also allow us to run them in parallel, which
should reduce the overall test duration (note that there are some
concerns around Cilium though).