Use self-hosted runners
We dropped the "session_instance" fixture, always using isolated
environments to prevent the test from interfering with each other.

The problem is that we're now running out of disk space, for which
reason we'll use large self-hosted runners. Most probably we were
already close to the node capacity, so this problem would've
surfaced sooner or later anyway.

We tried to clean up disk space using the same helper as the
Rockcraft jobs but it did not work, we sometime lose node connectivity:

  https://github.com/canonical/k8s-workflows/blob/main/.github/workflows/build_rocks.yaml#L200-L209

  The hosted runner: GitHub Actions 61 lost communication with the
  server. Anything in your workflow that terminates the runner
  process, starves it for CPU/Memory, or blocks its network access
  can cause this error.