Update "instances" fixture, removing the k8s snap when cleaning up
At the moment, the "instances" fixture removes the test instances
without uninstalling the k8s snap. In case of LXD instances, this
can leak certain resources on the host, for example loopback
devices.

We'll update the fixture to remove the k8s snap as part of the
instance cleanup procedure.