Updates the containerd-related paths
We're updating the default paths for containerd to align with the
default paths used by upstream. This makes sure any application that
requires interaction with containerd, such as the GPU operator can be
deployed with upstream instructions. However, we're only doing this for
classic confinement.

Adds ContainerdSocketPath, making it easier to refer to this path and
simplifying tests.

Adds a check in PreInitChecks, which ensures that there isn't a
containerd.sock file already existent in this location, meaning that
there is already a containerd instance running. This method is called in
the bootstrap and join hooks.

Adds unit tests.

[UpgradeImpact]: The default locations for the containerd-related files
has been changed for classic confinement snaps to align with the default
paths used by upstream, instead of the snap's common directory.