Retry removing k8s-snap and increase timeouts
We're now removing the k8s-snap before cleaning up the harness
instances, however it sometimes fails with:

  error: snap "k8s" has "service-control" change in progress

For this reason, we'll add some retries.

At the same time, the "strict" job fails while waiting for some of
the processes to go away. We'll try increasing the timeouts,
although this might be related to the "strict" job issues mentioned
by an inline comment.