Adds configurable containerd base dir during bootstrap / join
Currently, for the classic k8s snap, the default locations for the
containerd-related files are:

- /etc/containerd/
- /run/containerd/
- /var/lib/containerd/

These paths can conflict with other containerd installations on the host
(e.g. from docker), meaning that the k8s snap cannot be installed. In a
developer's usecase, we shouldn't require the other services to be
disabled, but we can allow the k8s snap's containerd to be installed in
a different location during bootstrap / node join.

Adds the --containerd-base-dir option for k8s bootstrap and join-cluster
commands. If not given, the snap's default paths will be used instead.
After the node was initialized, save the containerd base location in a
file, so we can properly reference it later / on restart.