Cleans up all containerd-related resources on snap removal
Currently, when removing the snap, the /var/run/containerd folder is not
properly removed, as it is a folder. This fixes this issue.

Additionally removes other containerd-related folders: /etc/containerd
and /var/lib/containerd.

We're also removing /opt/cni/bin on snap removal, which is created when
bootstrapping the node. As we're removing the k8s snap, we no longer
need this folder either.