Pick up doc update
Some docs are generated based on the k8s-snap-api docstrings.

We've just bumped the k8s-snap-api version and the docs are out of
sync because of a spelling fix. To fix the docs CI job, we need
to apply this fix here as well.

In the future, we may want to remove automatically generated docs
from the Github repo.