Make cilium cni.exclusive configurable (#838)
* Make cilium cni.exclusive configurable

We'll add an annotation, making the cilium "cni.exclusive" setting
configurable. This will allow using additional CNIs such as Multus.

* Address comments

* use "cni.exclusive=false" by default
* only check if the annotation is set, do not expect an actual
  boolean value
* s/AnnotationCniExclusive/AnnotationCNIExclusive

* bump k8s-snap-api to v1.0.14

* Pick up doc update

Some docs are generated based on the k8s-snap-api docstrings.

We've just bumped the k8s-snap-api version and the docs are out of
sync because of a spelling fix. To fix the docs CI job, we need
to apply this fix here as well.

In the future, we may want to remove automatically generated docs
from the Github repo.

* fix test assertion