Set default k8s snap track for registry (#852)
The registry setup fails if we do not set config.SNAP.
However, there a valid use cases, e.g. promotion tests in which a snap is not set.
We don't really care about the concrete snap version that is deployed on the registry as it is only used to collect inspection reports.