Checks that the expected services are running on bootstrap / join
Currently, we're doing only a few checks that would ensure a successful
installation of the k8s-snap (the containerd path is free for us to use,
the ports meant for Kubernetes services are free, binaries are sane).

However, due to various reasons, the bootstrap / join process can still
succeed, even if some of the services are not running (e.g.: invalid
extra service arguments given in the bootstrap config, leading to a
service to fail due to an unknown or invalid argument). At the moment,
we're only verifying that kube-apiserver is accessible.

We're adding a few checks that the Kubernetes services are not stopped /
failed after we've started them.

Adds integration test for this scenario.