Fix nightly tests (#876)
* Fix nightly job

The runner labels used by the nightly job are not correctly
evaluated.

We'll fix this by switching to the newly added one-string node
labels.

* Apply Docker iptables workaround

* Fix test_node_cleanup_new_containerd_path test

The containerd base dir setting is currently ignored when joining
nodes. The reason is that yaml parser ignores a struct field which
is marked as private. Making it public fixes the issue.

At the same time, the test expects the default containerd folders
to be missing when configured to use a different base dir.
However, the tests are currently placing the registry settings
in /etc/containerd. We also get an empty /run/containerd folder.
For now, we'll update the test to ignore these folders.

There's also an os.path.join call that doesn't produce the expected
result if the last element is an absolute path, we'll need to remove
a slash when constructing the paths.

* Fix linter errors

* Run all tests for backports or when there are test changes

We're adding a few checks to the integration tests workflow. We'll
use the "weekly" tag in case of backports (targeting the release-*
branches) and whenever there are test changes.

* tests: Add containerd cfg dir marker

One of the tests uses a custom containerd base dir. We'll add a
marker so that the registry mirror settings can be applied to that
containerd folder.

* Add yaml parser note