integration: fix test_containerd_path_cleanup_on_failed_init.
Previously, the test was inducing the `bootstrap` failure by attempting
to pre-allocate the port of the kube-controller-manager.

This methodology wasn't applicable on any non-LocalHarness testing
setups, so the test was updated to induce the failure by pre-creating
the containerd socket directory (`/run/containerd`), which should lead
to the same cleanup logic being triggered.

Signed-off-by: Nashwan Azhari <nashwan.azhari@canonical.com>