integration: completely remove LocalHarness.
The LocalHarness is a harness used for running integration tests
on the local machine where the tests are directly invoked
(be it via `tox`, `pytest`, etc).

It presents numerous limitations (can't run any multi-node tests)
and poses a lot of potential risks (cleanup failing in case of
fatal errors in the test fixtures) which outweighs most of its
convenience benefits (especially when compared to the LXD
substrate).

This patch completely removes the LocalHarness and all references to it
from the documentation, making LXD the new default.

Signed-off-by: Nashwan Azhari <nashwan.azhari@canonical.com>