Deduplicate github actions (#892)
* Deduplicate github actions

We have multiple github actions that run e2e tests and share a
significant amount of logic.

We'll add reusable actions, making the workflows much easier to
maintain.

* Fix flaky microk8s test

As part of the test cleanup, we're removing the k8s snap, ensuring
that its services and mounts go away.

One of the tests installs microk8s, which interferes with the k8s
snap cleanup assertions.

We'll fix this flaky test by removing the microk8s snap.

* Fix flaky ingress test

get_external_service_ip returns an empty string, however
the test asserts that the ip is not None and proceeds with the
curl:

  2024-12-12 11:28:46 DEBUG Execute command ['curl', '', '-H', 'Host: foo.bar.com']
  in instance k8s-integration-530bc4-37

We'll update the assertion and catch empty strings as well.

At the same time, we'll increase the timeouts to reduce test
flakiness.

* Merge nightly test and cron job

The nightly job is also a cron job that executes daily, so it
makes sense to merge those two workflows.

* Fix nightly job tag

* Pass test flavor

* Include all namespaces in inspection reports

The moonray job is failing, however we only have logs from the
"default" and "kube-system" namespaces.

This change will collect logs from all k8s namespaces.

* Apply flavor patches before running the tests

We'll need to apply the strict/moonray patches not only when
building the snap, but also when running the tests.

* Skip broken test

test_containerd_path_cleanup_on_failed_init holds an open port
and expects the bootstrap to fail, however that won't be the case
when using the lxd harness.

We'll skip this test for now.

* Revert "Include all namespaces in inspection reports"

This reverts commit 5020f3978648ad04cd989cf494ec77e938c4bc2a.

* Address PR feedback

* cover 1.32 as part of the nightly tests
* get go version from go.mod
* update step names
* add some TODOs
* make lxd channel configurable
* bump ubuntu versions
* add get-e2e-tags dependencies