test_ingress: fix service ip poller
get_external_service_ip repeatedly attempts to retrieve the
specified service ip.

The problem is that it uses "is None" checks to determine if
an ip was retrieved. This returns "False" for empty strings,
so the loop exits prematurely.

Note that "is None" checks are strongly discouraged in Python.

While at it, we're bumpting the timeout as well.