test_ingress: fix service ip poller (#956)
* test_ingress: fix service ip poller

get_external_service_ip repeatedly attempts to retrieve the
specified service ip.

The problem is that it uses "is None" checks to determine if
an ip was retrieved. This returns "False" for empty strings,
so the loop exits prematurely.

Note that "is None" checks are strongly discouraged in Python.

While at it, we're bumpting the timeout as well.

* Apply auto-generated docs changes

boostrap_config.md is auto-generated based on the k8s-snap-api
structures.

There's a job which ensures that the in-tree docs contain the latest
changes.

For now, we'll apply the expected changes. We should consider
removing the auto-generated docs from the git repo and add them to
.gitignore.