Don't remove Kubernetes certificates and containerd if skipped (#1002)
Currently, if the k8sd/v1alpha/lifecycle/skip-stop-services-on-remove
annotation is set, we're not stopping the Kubernetes-related services,
but we're still removing its certificates and containerd-related paths.
This will end up paralyzing services like kubelet, which might have to
do Pod evictions, blocking it from finishing its job, and resulting CAPI
not being able to complete its downscaling or upgrade operations.

We should remove those certificates only if we're also stopping the
services.