Fix pre-release update with same risk-level
The pre-release automation removes obsolete branches.
For that, it collects all obsolete tags and then removes the
corresponding branch.
However, we don't have patch level branches, so if the current release
is 1.33.0-alpha.1, the current implementation removed the
`autoupdate/1.33-alpha branch` since 1.33.0-alpha.0 is obsolete.

This fix now checks the risk level before removing the branch.