actions: Adds backport github action
Sets default action parameters explicitly.

In order to backport a PR, the PR should have the an appropriate label:
"backport branch-name" (e.g.: "backport 1.32"). When the PR merges, the
github action will automatically backport it. The label can also be
added after the PR merges, after which a user should comment /backport
in order to trigger the action.

The github-actions bot does not trigger any other CI workflows, but we'd
like to have tests running for the backports as well. Which is why the
backports will be made with a different account.

Additionally, we're now adding the "automerge" label to the new backports,
allowing the automerger action to merge the PR once the tests pass.