Adds non-default coredns ServiceAccount
The CIS 5.1.5 rule specifies: "Ensure that default service accounts are not actively used."

Currently, coredns is using the default service account. If this rule is
being enforced, coredns will no longer be able to function properly, as
it won't have the expected service account token mounted.

Adds an assertion that the coredns deployment is using the appropriate
service account in the integration tests.