Adds DrainConnectionsTimeout microcluster configuration (#867)
In microcluster, there is now a configuration that allows us to drain
any current connections before shutting down the unix / http servers.

Note that the servers will also be shut down when removing a node, or if
bootstrapping or joining a node fails. If we're not waiting for the
connection to drain, we may close current connections before managing to
write a response to the clients, resulting in them getting EOF errors.

Adds the --disable-update-node-config-controller k8sd flag, with 10s as
the default value.