Fix test_gateway retries (#1080)
test_gateway performs a few retries, waiting for the service IPs
to be set.

The problem is that the checks use "is None" syntax, which won't
properly handle empty strings. For this reason, the retry loop
stops prematurely, causing test flakiness.

   Execute command ['curl', '', '-H', 'Host: foo.bar.com']
   in instance k8s-integration-a6f840-37

In general "is None" / "is not None" checks should be avoided.
We'll update the k8s-snap tests accordingly.