tests: Adds Vault intermediate CA and external certificates integration tests (#1023)
* tests: Adds Vault intermediate CA integration test

The test sets up Vault in a harness instance, generates its root CA and
intermediate CAs and passes them to k8s-snap at bootstrap. The test
creates a cluster with 2 control plane nodes and a worker node.

* tests: Adds Vault external certificates test

The test sets up Vault in a harness instance, generates its root CA and
a few roles, and generates the certificates required by the Kubernetes nodes.
The test creates a cluster with 2 control plane nodes and a worker node.