bug: reject unknown bootstrap arguments (#1085)
* bug: reject unknown bootstrap arguments

We're updating the "k8s bootstrap" command to throw an error
when receiving unknown positional arguments.

For example, when passing a bootstrap config as a positional
argument instead of a named argument, the command will now error
out instead of ignoring the parameter, which can be confusing to
end users.

* use cobra.NoArgs for all the commands that don't take pos args