CI: use cdkbot token for automated jobs
The backport job creates the PRs using the cdkbot token but
approves them using "GITHUB_TOKEN". For this reason, the approval
is ignored and the backports cannot be merged automatically.

At the same time, the job that updates components uses the GH token
to create the PRs, which do get merged but the e2e tests are not
executed, potentially introducing regressions. Again, we'll have to
use the cdkbot token.