cilium: Enables protocol differentiation
Starting with cilium 1.17 [1][2], it now supports protocol
differentiation. Setting the loadBalancer.protocolDifferentiation.enabled
will result in ``bpf-lb-proto-diff`` being set in cilium's configmap,
which corresponds to this feature. We need this feature in order for the
following Conformance test to pass:

```
[sig-network] HostPort validates that there is no conflict between pods with same hostPort but different hostIP and protocol [LinuxOnly] [Conformance] [sig-network, Conformance]
```

However, that is not enough for the test to pass. We also need to
support `HostPort`s. According to the documentation [3], because we're
the standard kube-proxy and not cilium's replacement (``kubeProxyReplacement=false``),
we need to also set the Helm chart option ``cni.chainingMode=portmap``.
For the test above to pass, we also need to enable

[1] https://github.com/cilium/cilium/issues/9207
[2] https://github.com/cilium/cilium/pull/33434
[3] https://github.com/cilium/cilium/blob/v1.17.1/Documentation/installation/cni-chaining-portmap.rst#portmap-hostport