Fix: Temporarily disable snap/k8sd config sync during snap upgrade (#1164)
The snap and k8sd configurations are synchronized via k8s x-snapd-config in the configure hook. However, this process requires access to the k8sd API, which may be unavailable during an upgrade. For instance, in a clustered setup with microcluster, the upgrade process can be blocked until all nodes are on the same version, causing the reconciliation to time out.

We now explicitly disable the config sync during the snap upgrade to
avoid such scenarios.