ci: Changes PR linter event trigger
According to the documentation [1], ``pull_request_target`` events
should be used in a fork-based workflow.

If ``pull_request`` is being used intead, it will only work with PRs
from the same repository. With this configuration, if a PR is sent from
a fork, an error will occur.

Note that, according to the documentation, PRs that introduce this
GitHub action won't trigger the action (including this one), and PRs that
change the action's configuration will not be reflected in the updating
PRs.

[1] https://github.com/amannn/action-semantic-pull-request?tab=readme-ov-file#event-triggers