test: Adds test snap to test_version_upgrades
Currently, the test_version_upgrades test does not take into account the
configured TEST_SNAP (the built snap for testing the PR) when testing upgrades.
This means that if there's a PR that breaks upgrades, we won't notice
until the PR merges and gets published to the snap store, which is too
late.

This commit addresses this issue by adding the TEST_SNAP to the upgrade
test, helping ensure that the PR doesn't break upgrades.