fix: Execute post-refresh hook after node is ready
Initially, the custom post-refresh hook was executed in the microcluster
`onStart` hook. This hook is executed right after the microcluster
daemon started which could lead to "Database not initialized" error if
the node was not bootstrapped yet.

By introducing a nodeReadyHook and execute the refresh from there, we
ensure that the refresh is only executed after a `snap refresh` but not
after a `snap install`