CI: update the PR test job, dropping 20.04 and 22.04
At the moment, we're running the e2e tests on 20.04, 22.04 and 24.04.

To reduce the test runner load and the time it takes for a job to
be picked up, we'll only run 24.04 tests as part of the PR job.

Other versions and architectures (arm64) will be covered by the
nightly tests.

While at it, we'll also drop 20.04 from the nightly job.
It's going to hit EOL in two months and the charm tests dropped
20.04 as well.