test: Reduce CI log level to INFO
The integration test logs are overly verbose, especially in the CI where
we need to scroll to tons of logs to actually find the failing test.
This commit reduces the log level in the CI output to INFO and collects
the DEBUG level in a log file. If the job fails, the logs are uploaded
for further investigation.

The default (local) behavior is not changed, as one mostly runs only specific
tests locally for debugging.