test: Use multi-node cluster as part of upgrade/rollback tests (#1233)
At the moment, we have two almost identical checks that are usually
performed one after the other

* wait_until_k8s_ready
  * ensures that each instance node has "Ready" status
* ready_nodes followed by an assertion count
  * check each individual condition, expecting the the ones other than
    "Ready" to be "False"

This is redundant and we should merge those two checks. Furthermore,
it sometimes causes test flakiness as the first poller reports the
cluster as ready and then the second one fails.

Another problem is that "ready_nodes" doesn't log which condition
was not ready.

We'll improve those checks by:
* adding another helper called "is_node_ready", accepting either
  a node name or a node dict
* update wait_until_k8s_ready and ready_nodes to use the new helper
* drop unnecessary "ready_nodes" assertions after "wait_until_k8s_ready"
  since they're now redundant
  * the ready_nodes check is still useful when removing nodes

* Log disk usage and lxc containers before and after each test

We're currently running out disk space. To identify leaked
resources more easily, we'll log the disk usage and lxc containers
before and after each test.

* Ensure that we clean up even if we fail to collect insp reports