fix: Don't wait for services to terminate on removed node
Some services (e.g. k8s-dqlite) take a considerable amount of time to
shutdown in some cases. Microcluster has a fixed 30s timeout for the
remove hook which caused removing a node to fail in some scenarios.

With this commit, the remove hook will not wait anymore on services to
terminate. The cleanup on a removed node is on a best-effort basis so in
that case we don't care when (and even if) a service is actually stopped.