CI: allow invoking the e2e job manually
We'll add a "workflow_dispatch" to the e2e test job, allowing
it to be invoked manually through the GH actions UI.

By doing so, we'll able to run the e2e tests using GH actions
without having to send a PR or wait for the snap to build.

At the same time, we're adding another parameter for extra
pytest arguments. This can be used for example to run a specific
test (e.g. "-k <test_name>").