test: Skips the flaky CNCF test (#1261)
The test "validates resource limits of pods that are allowed to run [Conformance]"
currently fails randomly due to a scheduling error. If a node is
labeled, and immediately afterwards a Pod targeting that node label
is created, the Pod has a chance of failing to be scheduled properly,
causing the test to fail.

We're skipping the test until this issue is addressed.