test: enable worker refresh tests (#1255)
We had to disable worker refresh tests due to a known issue.
Once the fix gets backported and promoted to all the tracks
covered by the refresh tests, we can test worker nodes as well.

Follow-up of: https://github.com/canonical/k8s-snap/pull/1233
Requires the following backport to be merged and promoted:
https://github.com/canonical/k8s-snap/pull/1239