Revert "test: enable worker refresh tests (#1255)"
This reverts commit 2a3d7b1c810f3ca1d24a5ea4584b8008d016a553.

The arm64 revisions seem to be out of sync with the amd64 ones.
Let's revert this for now to unblock the nightly CI job and send
it again in a couple of weeks, by that time the fix should reach all
the snap channels.

Important: run the nightly tests before merging this back.