test: properly wait for k8sd controllers to finish reconciling (#1273)
The k8sd controller unit tests will emit a fake event and wait
for the reconcile loop to handle it.

UpdateNodeConfigurationController uses a channel to notify that
the reconciliation finished, however the other tests will just
wait for about 100ms and expect the reconcile to finish.

The 100ms sleep is flaky, sometimes it takes longer for the events
to be processed.

To address this, we'll use the same post-reconciliation notification
in all k8sd controllers.