Avoid spawning harness instances for xfail tests
We have a few tests that are currently skipped using pytest.xfail.

The problem is that the harness instances are initialized through
fixtures before pytest.xfail gets called, so we waste time and
resources by spawning harness instances for tests that are skipped
anyway.

This can be avoided by using the pytest.mark.xfail decorator, which
gets handled before the "instances" fixture.