Backport 1228 and 1262 (#1287)
* fix: Execute post-refresh hook after node is ready (#1228)

Initially, the custom post-refresh hook was executed in the microcluster
`onStart` hook. This hook is executed right after the microcluster
daemon started which could lead to "Database not initialized" error if
the node was not bootstrapped yet.

By introducing a nodeReadyHook and execute the refresh from there, we
ensure that the refresh is only executed after a `snap refresh` but not
after a `snap install`

* fix: Don't wait for services to terminate on removed node (#1262)

Some services (e.g. k8s-dqlite) take a considerable amount of time to
shutdown in some cases. Microcluster has a fixed 30s timeout for the
remove hook which caused removing a node to fail in some scenarios.

With this commit, the remove hook will not wait anymore on services to
terminate. The cleanup on a removed node is on a best-effort basis so in
that case we don't care when (and even if) a service is actually stopped.

* test: small e2e test improvements (#1279)

---------

Co-authored-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>