test: ensure that the containerd test uses local mirror
One of the tests uses a custom containerd base directory. The
problem is that it places the registry configuration at a wrong
location that gets ignored.

For this reason, this test will always try to pull the images
from the upstream registry, which causes flakiness (timeouts
due to the image size and api rate limiting).

We'll solve the issue by using the same directory that is
specified in the containerd config file:

```
[plugins."io.containerd.grpc.v1.cri".registry]
  config_path = "/home/ubuntu/k8s-containerd/etc/containerd/hosts.d"
```