test: generate subunit test reports
The test logs are quite verbose (which is good), however it's
difficult to find the results of a given test that failed.

To address this, we'll stream the test results in subunit
format[1][2][3], which is heavily used by Openstack projects.

The pytest-subunit [4][5] plugin seems unmaintained, however it's
lightweight and can be simplified even further.

For now, I've created a fork [6] that fixes the plugin and modifies it
to use a separate subunit file instead of streaming subunit data
directly on stdout.

TODO:

* get some feedback on the html reports
* decide where to put the pytest extension (separate repo or inside the
  test dir)

[1] https://pypi.org/project/python-subunit/
[2] https://launchpad.net/subunit
[3] https://github.com/testing-cabal/subunit

[4] https://pypi.org/project/pytest-subunit/
[5] https://github.com/jelmer/pytest-subunit

[6] https://github.com/petrutlucian94/pytest-subunit